refactor: inline software identifier string rules
